ft: switch to using an abstract base class
