+++ vCardLib/Models/vCard.cs
@@ -194,7 +194,7 @@ namespace vCardLib.Models
             switch (version ?? Version)
             {
                 case vCardVersion.V2:
-                    return new V2Serializer().Serialize(this);
+                    return new v2Serializer().Serialize(this);
                 case vCardVersion.V3:
                     return new V3Serializer().Serialize(this);
                 default:
