+++ vCardLib/Serializers/V4Serializer.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Text;
-using vCardLib.Interfaces;
 using vCardLib.Models;
 
 namespace vCardLib.Serializers
@@ -9,54 +8,44 @@ namespace vCardLib.Serializers
     /// <summary>
     /// Handles the serialization of version 4 cards
     /// </summary>
-    internal class V4Serializer : BaseSerializer, ISerializer
+    internal class V4Serializer : BaseSerializer
     {
-        public string Serialize(vCard vCard)
+        protected override void AddVersion(StringBuilder stringBuilder)
         {
             throw new NotImplementedException();
         }
 
-        public string Serialize(IEnumerable<vCard> vCardCollection)
+        protected override void AddPhoneNumbers(StringBuilder stringBuilder, IEnumerable<PhoneNumber> phoneNumbers)
         {
             throw new NotImplementedException();
         }
 
-        public void AddVersion(StringBuilder stringBuilder)
-        {
-            stringBuilder.AppendLine("VERSION:4.0");
-        }
-
-        public void AddPhoneNumbers(StringBuilder stringBuilder, IEnumerable<PhoneNumber> phoneNumbers)
+        protected override void AddEmailAddresses(StringBuilder stringBuilder, IEnumerable<EmailAddress> emailAddresses)
         {
             throw new NotImplementedException();
         }
 
-        public void AddEmailAddresses(StringBuilder stringBuilder, IEnumerable<EmailAddress> emailAddresses)
+        protected override void AddAddresses(StringBuilder stringBuilder, IEnumerable<Address> addresses)
         {
             throw new NotImplementedException();
         }
 
-        public void AddAddresses(StringBuilder stringBuilder, IEnumerable<Address> addresses)
+        protected override void AddPhotos(StringBuilder stringBuilder, IEnumerable<Photo> photos)
         {
             throw new NotImplementedException();
         }
 
-        public void AddPhotos(StringBuilder stringBuilder, IEnumerable<Photo> photos)
+        protected override void AddExpertises(StringBuilder stringBuilder, IEnumerable<Expertise> expertises)
         {
             throw new NotImplementedException();
         }
 
-        public void AddExpertises(StringBuilder stringBuilder, IEnumerable<Expertise> expertises)
+        protected override void AddHobbies(StringBuilder stringBuilder, IEnumerable<Hobby> hobbies)
         {
             throw new NotImplementedException();
         }
 
-        public void AddHobbies(StringBuilder stringBuilder, IEnumerable<Hobby> hobbies)
-        {
-            throw new NotImplementedException();
-        }
-
-        public void AddInterests(StringBuilder stringBuilder, IEnumerable<Interest> interests)
+        protected override void AddInterests(StringBuilder stringBuilder, IEnumerable<Interest> interests)
         {
             throw new NotImplementedException();
         }
